Patience plus added to project + datatype changed
datatype of values which need to be sorted are changed from int to long
long. qsort doesn't seem to like this for some reason. Need to look in
to that

+++ sorting/patience.cpp
@@ -56,7 +56,7 @@ void patience_sort(Iterator begin, Iterator end)
 	}
 }
 
-int patsort(int values[], int length) {
+int patsort(long long values[], int length) {
 	patience_sort(values, values + length);
 	return 0;
 }
